Update ownership check, overlaps

+++ controllers/timeSplits.ts
@@ -8,32 +8,28 @@ import {
 import log from "loglevel";
 
 const create_time_splits = asyncHandler(async (req, res, next) => {
+  await checkContentOwnership(req, res, next);
+
   const newSplitsForDb = await validateTimeSplitRequest(req, res, next);
 
-  try {
-    const splits = await prisma.timeSplit.createMany({
-      data: newSplitsForDb,
-    });
+  if (newSplitsForDb) {
+    try {
+      const splits = await prisma.timeSplit.createMany({
+        data: newSplitsForDb,
+      });
 
-    res.status(200).json({ success: true, data: splits });
-  } catch (e) {
-    const error = formatError(500, `${e.code}: ${e.message}`);
-    next(error);
-    return;
+      res.status(200).json({ success: true, data: splits });
+    } catch (e) {
+      const error = formatError(500, `${e.code}: ${e.message}`);
+      next(error);
+      return;
+    }
   }
 });
 
 const get_time_splits = asyncHandler(async (req, res, next) => {
-  const { contentId } = req.body;
-  // Check where clause
-  if (!contentId) {
-    const error = formatError(400, "contentId is required");
-    next(error);
-    return;
-  }
-
-  // Does user own this content?
   await checkContentOwnership(req, res, next);
+  const { contentId } = req.body;
 
   try {
     const splits = await prisma.timeSplit.findMany({
@@ -51,13 +47,8 @@ const get_time_splits = asyncHandler(async (req, res, next) => {
 });
 
 const update_time_splits = asyncHandler(async (req, res, next) => {
+  await checkContentOwnership(req, res, next);
   const { contentId } = req.body;
-  // Check where clause
-  if (!contentId) {
-    const error = formatError(400, "contentId is required");
-    next(error);
-    return;
-  }
 
   // Delete splits if time split list is empty
   const timeSplits = req.body.timeSplits;
@@ -83,26 +74,28 @@ const update_time_splits = asyncHandler(async (req, res, next) => {
 
   const newSplitsForDb = await validateTimeSplitRequest(req, res, next, true);
 
-  try {
-    const splits = await prisma.timeSplit
-      .deleteMany({
-        where: {
-          contentId: {
-            equals: contentId,
+  if (newSplitsForDb) {
+    try {
+      const splits = await prisma.timeSplit
+        .deleteMany({
+          where: {
+            contentId: {
+              equals: contentId,
+            },
           },
-        },
-      })
-      .then(() => {
-        return prisma.timeSplit.createMany({
-          data: newSplitsForDb,
+        })
+        .then(() => {
+          return prisma.timeSplit.createMany({
+            data: newSplitsForDb,
+          });
         });
-      });
 
-    res.status(200).json({ success: true, data: splits });
-  } catch (e) {
-    const error = formatError(500, `${e.code}: ${e.message}`);
-    next(error);
-    return;
+      res.status(200).json({ success: true, data: splits });
+    } catch (e) {
+      const error = formatError(500, `${e.code}: ${e.message}`);
+      next(error);
+      return;
+    }
   }
 });
